+++ api-types/src/lib.rs
@@ -17,9 +17,11 @@ pub use migration::{
     MigrationMode, TimeoutStrategy, VmReceiveMigrationData, VmReceiveMigrationDataParseError,
     VmSendMigrationData, VmSendMigrationDataParseError,
 };
-pub use restore::{RestoredVfioConfig, VmMemoryZoneUpdateData};
+pub use restore::{
+    MemoryRestoreMode, MemoryRestoreModeParseError, RestoredNetConfig, RestoredVfioConfig,
+    VmMemoryZoneUpdateData,
+};
 use serde::Deserialize;
-pub use restore::{MemoryRestoreMode, MemoryRestoreModeParseError};
 pub use vm_coredump::VmCoredumpData;
 pub use vm_remove_device::VmRemoveDeviceData;
 pub use vm_resize::VmResizeData;
